Fixed DynamoDb logic & Added Test conainer for it
