+++ ManagedCode.Database.DynamoDB/DynamoDBDatabase.cs
@@ -1,5 +1,9 @@
 ﻿using Amazon.DynamoDBv2;
+using Amazon.DynamoDBv2.DataModel;
+using Amazon.DynamoDBv2.DocumentModel;
+using Amazon.DynamoDBv2.Model;
 using ManagedCode.Database.Core;
+using ManagedCode.Database.Core.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,35 +11,47 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace ManagedCode.Database.DynamoDB
+namespace ManagedCode.Database.DynamoDB;
+
+public class DynamoDBDatabase : BaseDatabase<AmazonDynamoDBClient>
 {
-    internal class DynamoDBDatabase : BaseDatabase<AmazonDynamoDBClient>
+    private readonly Dictionary<string, AttributeValue> _collections = new();
+
+    private readonly DynamoDBOptions _dbOptions;
+    private DynamoDBContext _dynamoDBContext;
+
+    public DynamoDBDatabase(DynamoDBOptions dbOptions)
     {
-        private readonly DynamoDBOptions _dbOptions;
+        _dbOptions = dbOptions;
+    }
 
-        public DynamoDBDatabase(DynamoDBOptions dbOptions)
-        {
-            _dbOptions = dbOptions;
-        }
+    public override Task DeleteAsync(CancellationToken token = default)
+    {
+        throw new NotImplementedException();
+    }
 
-        public override Task DeleteAsync(CancellationToken token = default)
-        {
-            throw new NotImplementedException();
-        }
+    protected override Task InitializeAsyncInternal(CancellationToken token = default)
+    {
+        var client = new AmazonDynamoDBClient();
+        _dynamoDBContext = new DynamoDBContext(client);
 
-        protected override ValueTask DisposeAsyncInternal()
-        {
-            throw new NotImplementedException();
-        }
+        return Task.CompletedTask;
+    }
 
-        protected override void DisposeInternal()
-        {
-            throw new NotImplementedException();
-        }
+    protected override ValueTask DisposeAsyncInternal()
+    {
+        return new ValueTask(Task.CompletedTask);
+    }
 
-        protected override Task InitializeAsyncInternal(CancellationToken token = default)
-        {
-            throw new NotImplementedException();
-        }
+    protected override void DisposeInternal()
+    {
+    }
+
+    public DynamoDBCollection<TItem> GetCollection<TItem>()
+       where TItem : DynamoDBItem, new()
+    {
+        if (!IsInitialized) throw new DatabaseNotInitializedException(GetType());
+
+        return new DynamoDBCollection<TItem>(_dynamoDBContext);
     }
 }
