+++ ManagedCode.Database.DynamoDB/DynamoDBOptions.cs
@@ -4,12 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ManagedCode.Database.DynamoDB
+namespace ManagedCode.Database.DynamoDB;
+
+public class DynamoDBOptions
 {
-    internal class DynamoDBOptions
-    {
-        public string ConnectionString { get; set; }
-        public string DataBaseName { get; set; }
-        public string AllowTableCreation { get; set; }
-    }
+    public string ConnectionString { get; set; }
+    public string DataBaseName { get; set; }
+    //public bool AllowTableCreation { get; set; }
 }
