+++ ManagedCode.Database.Tests/ManagedCode.Database.Tests.csproj
@@ -51,6 +51,7 @@
         <ProjectReference Include="..\ManagedCode.Database.AzureTables\ManagedCode.Database.AzureTables.csproj" />
         <ProjectReference Include="..\ManagedCode.Database.Core\ManagedCode.Database.Core.csproj" />
         <ProjectReference Include="..\ManagedCode.Database.Cosmos\ManagedCode.Database.Cosmos.csproj" />
+        <ProjectReference Include="..\ManagedCode.Database.DynamoDB\ManagedCode.Database.DynamoDB.csproj" />
         <ProjectReference Include="..\ManagedCode.Database.LiteDB\ManagedCode.Database.LiteDB.csproj" />
         <ProjectReference Include="..\ManagedCode.Database.MongoDB\ManagedCode.Database.MongoDB.csproj" />
         <ProjectReference Include="..\ManagedCode.Database.SQLite\ManagedCode.Database.SQLite.csproj" />
